+++ app/routes/Icons.js
@@ -4,8 +4,6 @@
 
 import {CodeBlock, Icon, ICON_ROUND_ADD} from '../../src'
 import route from './factory'
-// $FlowFixMe - Flow doesn't understand SASS imports
-import './Icons.scss'
 import React, {Component, type Node} from 'react'
 
 const BACKGROUND_COLORS = [