+++ plugins/babel-plugin-auto-css-modules.js
@@ -1,0 +1,22 @@
+const {existsSync} = require('fs')
+const {basename, dirname, join} = require('path')
+
+module.exports = ({types: t, template}) => {
+  return {
+    visitor: {
+      Program(path, state) {
+        const {filename} = state.file.opts
+        const fileNameWithoutExtension = basename(filename, '.js')
+        const directory = dirname(filename)
+        const cssModulePath = join(directory, `${fileNameWithoutExtension}.css`)
+        const cssModuleExists = existsSync(cssModulePath)
+
+        if (cssModuleExists) {
+          path.node.body.unshift(
+            t.importDeclaration([], t.stringLiteral(`./${fileNameWithoutExtension}.css`))
+          )
+        }
+      }
+    },
+  }
+}