+++ src/components/__tests__/Text-test.js
@@ -1,0 +1,9 @@
+import Text from '../Text'
+import {mount} from 'enzyme'
+import React from 'react'
+
+describe('Text', () => {
+  it('renders as expected without any properties passed in', () => {
+    expect(mount(<Text />)).toMatchSnapshot()
+  })
+})