+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+*   Switched from SASS to postcss/cssnext.
+*   Auto-generating color background classes and font size classes.
+
 ## 0.0.2
 
 *   Added text input component.